+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/preferences/GoSDKConfigBlock.java
@@ -11,7 +11,6 @@
 package com.googlecode.goclipse.ui.preferences;
 
 import static melnorme.lang.tooling.ops.util.PathValidator.LocationKind.FILE_ONLY;
-import static melnorme.utilbox.core.CoreUtil.array;
 
 import java.io.File;
 
@@ -25,8 +24,6 @@ import org.eclipse.swt.widgets.Group;
 
 import com.googlecode.goclipse.core.GoEnvironmentPrefs;
 import com.googlecode.goclipse.tooling.GoSDKLocationValidator;
-import com.googlecode.goclipse.tooling.env.GoArch;
-import com.googlecode.goclipse.tooling.env.GoOs;
 
 import melnorme.lang.ide.ui.preferences.AbstractCompositePreferencesBlock;
 import melnorme.lang.ide.ui.preferences.common.AbstractPreferencesBlock2;
@@ -34,15 +31,12 @@ import melnorme.lang.ide.ui.preferences.common.PreferencesPageContext;
 import melnorme.lang.tooling.data.IValidatableValue.ValidatableField;
 import melnorme.lang.tooling.ops.util.LocationValidator;
 import melnorme.lang.utils.EnvUtils;
-import melnorme.util.swt.SWTFactoryUtil;
 import melnorme.util.swt.components.fields.CheckBoxField;
-import melnorme.util.swt.components.fields.ComboBoxField;
 import melnorme.util.swt.components.fields.DirectoryTextField;
 import melnorme.util.swt.components.fields.EnablementButtonTextField;
 import melnorme.util.swt.components.fields.FileTextField;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
-import melnorme.utilbox.misc.ArrayUtil;
 import melnorme.utilbox.misc.MiscUtil;
 
 public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
@@ -54,13 +48,6 @@ public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
 	protected final FileTextField goFmtPath = new FileTextField("gofmt:");
 	protected final FileTextField goDocPath = new FileTextField("godoc:");
 	
-	protected final ComboBoxField goOSField = new ComboBoxField("G&OOS:",
-		ArrayUtil.prepend("<default>", GoOs.GOOS_VALUES),
-		ArrayUtil.prepend("", GoOs.GOOS_VALUES));
-	protected final ComboBoxField goArchField = new ComboBoxField("GO&ARCH:", 
-		array("<default>", GoArch.ARCH_AMD64, GoArch.ARCH_386,  GoArch.ARCH_ARM), 
-		array(""         , GoArch.ARCH_AMD64, GoArch.ARCH_386, GoArch.ARCH_ARM));
-	
 	protected final EnablementButtonTextField goPathField = new GoPathField();
 	protected final CheckBoxField gopathAppendProjectLocField = new CheckBoxField(
 		"Also add project location to GOPATH, if it's not contained there already.");
@@ -71,8 +58,6 @@ public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
 		addSubComponent(goRootField);
 		addSubComponent(goFmtPath);
 		addSubComponent(goDocPath);
-		addSubComponent(goOSField);
-		addSubComponent(goArchField);
 		
 		addSubComponent(goPathField);
 		addSubComponent(gopathAppendProjectLocField);
@@ -80,9 +65,6 @@ public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
 		prefContext.bindToPreference(goRootField, GoEnvironmentPrefs.GO_ROOT);
 		prefContext.bindToPreference(goFmtPath, GoEnvironmentPrefs.FORMATTER_PATH);
 		prefContext.bindToPreference(goDocPath, GoEnvironmentPrefs.DOCUMENTOR_PATH);
-		
-		prefContext.bindToPreference(goOSField.asStringProperty(), GoEnvironmentPrefs.GO_OS);
-		prefContext.bindToPreference(goArchField.asStringProperty(), GoEnvironmentPrefs.GO_ARCH);
 		
 		prefContext.bindToPreference(goPathField.asEffectiveValueProperty2(), GoEnvironmentPrefs.GO_PATH);
 		prefContext.bindToPreference(gopathAppendProjectLocField, GoEnvironmentPrefs.APPEND_PROJECT_LOC_TO_GOPATH);
@@ -114,13 +96,6 @@ public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
 		
 		goFmtPath.createComponentInlined(goSDK);
 		goDocPath.createComponentInlined(goSDK);
-		
-		SWTFactoryUtil.createLabel(goSDK, 
-			SWT.SEPARATOR | SWT.HORIZONTAL, "",
-			gdFillDefaults().span(numColumns, 1).grab(true, false).indent(0, 5).create());
-		
-		goOSField.createComponentInlined(goSDK);
-		goArchField.createComponentInlined(goSDK);
 		
 		/* -----------------  ----------------- */
 		
@@ -160,7 +135,7 @@ public class GoSDKConfigBlock extends AbstractCompositePreferencesBlock {
 	public static class GoPathField extends EnablementButtonTextField {
 		
 		public GoPathField() {
-			super("GOPATH:", "Use same as GOPATH environment variable.", "Add Folder");
+			super("Eclipse GOPATH:", "Use same as the GOPATH environment variable.", "Add Folder");
 		}
 		
 		@Override