+++ plugin_tooling/src/com/googlecode/goclipse/tooling/oracle/GoOracleDescribeOperation.java
@@ -39,11 +39,12 @@ public class GoOracleDescribeOperation extends ToolOutputParseHelper {
 		}
 		
 		// BM: I thought go oracle required these variables to be set, but it seems its not the case anymore?
-		boolean requireNonNullGoOSGoArch = false;
-		if(requireNonNullGoOSGoArch) {
-			goEnv.getGoArch_NonNull();
-			goEnv.getGoOs_NonNull();
-		}
+		// So commented them out
+//		boolean requireNonNullGoOSGoArch = false;
+//		if(requireNonNullGoOSGoArch) {
+//			goEnv.getGoArch_NonNull();
+//			goEnv.getGoOs_NonNull();
+//		}
 		
 		String goPackageString = goPackage.getFullNameAsString();